UI update, missions
Added unit summaries and unit info commands (still work to be done).
Battlefields are now represented by two files: a map.json file with info about the board itself (generated from MeshToBoard), and a mission.json file which is responsible for setting up units (and in the future, event flags).