Improved movement
Units now move with an animation, and automatically choose a path to do so with as few sharp turns as possible (in a way that looks "natural"). Universal model and animator controller are used, and will be extended for attack animations. Speed-up of movement animations while a certain button is held will be implemented.